+++ juniper/src/types/async_await.rs
@@ -260,6 +260,16 @@ where
                         .await;
 
                     let value = match res {
+                        // This hack is required as Juniper doesn't allow at the
+                        // moment for custom defined types to tweak into
+                        // resolving.
+                        // TODO: Redesign resolving layer to allow such things.
+                        #[cfg(feature = "json")]
+                        Ok(Value::Null)
+                            if is_non_null && meta_field.field_type.name() == Some("Json") =>
+                        {
+                            Some(Value::Null)
+                        }
                         Ok(Value::Null) if is_non_null => None,
                         Ok(v) => Some(v),
                         Err(e) => {